panels: firmware-security: Host security ID panel for fwupd
This commit is used to display the host security levels and details to reveal the
security status to the user. The information is generated by fwupd. This panel
allows the user immediately aware of the harware securty level by displaying some
colourful large icons. The panel also reveals the hardware configuration changes
to pinpoint the configuration changing time.